教學專案/langchain/requirements.txt
# LangChain 教學專案依賴套件

# 核心 LangChain 套件
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# 文件處理套件
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# 向量資料庫套件
chromadb>=0.4.0
faiss-cpu>=1.7.0
pinecone-client>=2.2.0

# 資料處理
pandas>=2.0.0
numpy>=1.24.0

# 環境變數管理
python-dotenv>=1.0.0

# 視覺化
matplotlib>=3.8.0
seaborn>=0.13.0

# 進階功能
langchain-experimental>=0.0.50
langgraph>=0.1.0

# 工具整合
wikipedia>=1.4.0
duckduckgo-search>=3.9.0
python-multipart>=0.0.6

# 測試和開發
pytest>=7.0.0
jupyter>=1.0.0